add: Layout: NavLinks
